refactor: move phone utilities to a separate file and update imports in schemas and tests
